extract metrics utils
