+++ query-engine/connector-test-kit-rs/query-engine-tests/Cargo.toml
@@ -27,3 +27,4 @@ paste = "1.0.14"
 
 [dev-dependencies]
 insta = "1.7.1"
+itertools.workspace = true
